LayoutTests/css3/flexbox/flex-flow-initial.html

<!DOCTYPE html>
<style>
#test1 {
  flex-flow: column;
}
#test1 {
  flex-flow: initial;
}
#test2 {
  flex-direction: column;
}
#test2 {
  flex-direction: initial;
}
</style>
<p>Tests that 'initial' is handled correctly for the flex-flow shorthand.</p>
<div id="log"></div>
<span id="test1"></span>
<span id="test2"></span>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
value1 = window.getComputedStyle(test1).flexDirection;
value2 = window.getComputedStyle(test2).flexDirection;
test(function() {
assert_equals(window.getComputedStyle(test1).flexDirection, "row");
assert_equals(window.getComputedStyle(test2).flexDirection, "row");
}, "initial value is correct");
if (window.testRunner)
   testRunner.dumpAsText();
</script>
